requirements.txt
langchain-google-genai
python-dotenv
streamlit
PyPDF2
reportlab
markdown
beautifulsoup4